build: modularize development toolkit and implement rich logging
- Decomposed the monolithic `dev/main.py` into specialized modules: `env`, `qa`, `docs`, `infra`, and `release`.
- Implemented a centralized logging configuration in `dev/common.py` using `RichHandler` for styled, severity-aware CLI output.
- Added `rich` to the `dev` dependency group in `pyproject.toml`.
- Refined `dev/main.py` to act as a facade, preserving compatibility with existing `uv run` entry points.
- Added comprehensive docstrings across the development toolkit to improve maintainability and contributor onboarding.
- Standardized the `run_command` utility to provide a consistent UI experience across all development tasks.
